finish subsection: Lifting the State Up
