+++ src/App.js
@@ -1,25 +1,60 @@
-import logo from './logo.svg';
-import './App.css';
+import "./App.css";
+import Counters from "./components/counters";
+import NavBar from "./components/navbar";
+import React, { Component } from "react";
 
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
+class App extends Component {
+  state = {
+    counters: [
+      { id: 1, value: 0 },
+      { id: 2, value: 11 },
+      { id: 3, value: 21 },
+      { id: 4, value: 31 },
+    ],
+  };
+
+  handleReset = () => {
+    const counters = this.state.counters.map((c) => {
+      c.value = 0;
+      return c;
+    });
+    this.setState({ counters });
+  };
+
+  handleIncrement = (counter) => {
+    console.log("Incrementing value of counter with id: " + counter.id);
+    const counters = [...this.state.counters];
+    const index = counters.indexOf(counter);
+    counters[index] = { ...counter };
+    counters[index].value++;
+    this.setState({ counters });
+  };
+
+  handleDelete = (id) => {
+    console.log("Removing counter-id: " + id);
+    const counters = this.state.counters.filter((counter) => counter.id !== id);
+    this.setState({ counters });
+  };
+
+  render() {
+    return (
+      <React.Fragment>
+        <NavBar
+          countersCount={
+            this.state.counters.filter((c) => c.value !== 0).length
+          }
+        />
+        <main className="container">
+          <Counters
+            counters={this.state.counters}
+            onDelete={this.handleDelete}
+            onIncrement={this.handleIncrement}
+            onReset={this.handleReset}
+          />
+        </main>
+      </React.Fragment>
+    );
+  }
 }
 
 export default App;
