+++ src/components/counters.jsx
@@ -2,53 +2,21 @@ import React, { Component } from "react";
 import Counter from "./counter";
 
 class Counters extends Component {
-  state = {
-    counters: [
-      { id: 1, value: 0 },
-      { id: 2, value: 11 },
-      { id: 3, value: 21 },
-      { id: 4, value: 31 },
-    ],
-  };
-
-  handleReset = () => {
-    const counters = this.state.counters.map((c) => {
-      c.value = 0;
-      return c;
-    });
-    this.setState({ counters });
-  };
-
-  handleIncrement = (counter) => {
-    console.log("Incrementing value of counter with id: " + counter.id);
-    const counters = [...this.state.counters];
-    const index = counters.indexOf(counter);
-    counters[index] = { ...counter };
-    counters[index].value++;
-    this.setState({ counters });
-  };
-
-  handleDelete = (id) => {
-    console.log("Removing counter-id: " + id);
-    const counters = this.state.counters.filter((counter) => counter.id !== id);
-    this.setState({ counters });
-  };
-
   render() {
     return (
       <div>
         <button
-          onClick={this.handleReset}
+          onClick={this.props.onReset}
           className="btn btn-primary btn-sm m-2"
         >
           Reset
         </button>
-        {this.state.counters.map((c) => (
+        {this.props.counters.map((c) => (
           <Counter
             key={c.id}
             counter={c}
-            onDelete={this.handleDelete}
-            onIncrement={this.handleIncrement}
+            onDelete={this.props.onDelete}
+            onIncrement={this.props.onIncrement}
           />
         ))}
       </div>
